//1.3.4 & 1.3.5
creating flows and loading them, also changes on Accounts class so it
can receive a Flow object

+++ GANBank/src/components/Credit.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class Credit extends Flow{
 	//1.3.3 Creation of Credit class
-	public Credit(String comment, String identifier, int amount, int targetAccountNumber, boolean effect,
+	public Credit(String comment, String identifier, double amount, int targetAccountNumber, boolean effect,
 			LocalDate dateOfFlowDate) {
 		super(comment, identifier, amount, targetAccountNumber, effect, dateOfFlowDate);
 	}
