+++ GANBank/src/components/Debit.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class Debit  extends Flow{
 	//1.3.3 Creation of Debit class
-	public Debit(String comment, String identifier, int amount, int targetAccountNumber, boolean effect,
+	public Debit(String comment, String identifier, double amount, int targetAccountNumber, boolean effect,
 			LocalDate dateOfFlowDate) {
 		super(comment, identifier, amount, targetAccountNumber, effect, dateOfFlowDate);
 	}
